requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
gym
pybullet
mujoco==2.3.0
imageio
pyyaml